#016-Core: 04 August 2026: Added clang-tidy usage and fixed all new linter errors

+++ src/SunnahTimes.cpp
@@ -10,7 +10,8 @@ SunnahTimes::SunnahTimes(const PrayerTimes &prayerTimes) {
       prayerTimes.coordinates, nextDay, prayerTimes.calculationParameters);
 
   auto nightDuration =
-      (nextDayPrayerTimes.fajr.getTime() - prayerTimes.maghrib.getTime()) /
+      static_cast<double>(
+          nextDayPrayerTimes.fajr.getTime() - prayerTimes.maghrib.getTime()) /
       1000.0;
 
   this->middleOfTheNight = roundedMinute(
